refactor(http-server): extract in-flight request tracking into server/in_flight.rs
Moves the #4728 in-flight (async-handler) request tracking + reaper
cluster out of server.rs into a new sibling submodule
server/in_flight.rs: the InFlightRequest struct, the IN_FLIGHT static,
and response_writable_ended / finalize_request_handles /
finalize_request_handles_deferred / has_in_flight_requests /
reap_in_flight_requests / finalize_or_park_request.

This brings server.rs from 2025 to 1836 lines, back under the 2000-line
file-size limit enforced by scripts/check_file_size.sh (currently red on
main, blocking the lint gate). Pure move: the relocated bodies are
byte-identical apart from the pub(crate) visibility widening needed for
the cross-module references and the new module's use header. No behavior
change; perry-ext-http-server tests pass.
